ajust conflit

+++ src/components/multiply/multiply.go
@@ -1,7 +1,7 @@
 package multiply
 
 // Execute returns the product of the integers in the matrix
-func Execute(m [][]int) (s int) {
+func Execute(m [][]int64) (s int64) {
 	if len(m) == 0 {
 		return 0
 	}
